Move MCP contents into subfolder. Great idea.
