--- claude_desktop_config.json
@@ -1,9 +1,0 @@
-{
-  "mcpServers": {
-    "inat-diff": {
-      "command": "python3",
-      "args": ["-m", "mcp_server"],
-      "cwd": "/path/to/inat-diff"
-    }
-  }
-}
