+++ inat-diff-mcp/install_mcp.sh
@@ -249,8 +249,8 @@ echo "  3. Try asking Claude:"
 echo "     \"What new species appeared in Oregon this month?\""
 echo ""
 echo "For help and examples, see:"
+echo "  - QUICKSTART.md (quick start guide)"
 echo "  - MCP_README.md (detailed documentation)"
-echo "  - INSTALL_MCP.md (troubleshooting)"
 echo ""
 echo "Configuration file: $CONFIG_FILE"
 echo "Installation directory: $INSTALL_DIR"
