+++ inat-diff-mcp/requirements-mcp.txt
@@ -1,0 +1,4 @@
+# MCP Server dependencies
+# Install with: pip install -r requirements.txt -r requirements-mcp.txt
+
+mcp>=1.0.0
