+++ inat-diff-mcp/smithery.json
@@ -1,0 +1,36 @@
+{
+  "name": "inat-diff",
+  "displayName": "iNaturalist Difference Detection",
+  "description": "MCP server for invasive species monitoring and biodiversity research through iNaturalist data. Find new species in regions, track species presence patterns, and monitor biodiversity changes over time.",
+  "category": "science",
+  "tags": ["biodiversity", "ecology", "invasive-species", "inaturalist", "conservation", "biology"],
+  "homepage": "https://github.com/agentmorris/inat-diff",
+  "repository": "https://github.com/agentmorris/inat-diff",
+  "license": "MIT",
+  "author": {
+    "name": "iNaturalist Difference Detection Team"
+  },
+  "mcp": {
+    "command": "python3",
+    "args": ["-m", "mcp_server"],
+    "requiresConfig": false
+  },
+  "installation": {
+    "type": "pip",
+    "package": "inat-diff[mcp]",
+    "pipInstallCommand": "pip install 'inat-diff[mcp]'"
+  },
+  "requirements": {
+    "python": ">=3.10"
+  },
+  "keywords": [
+    "mcp",
+    "model-context-protocol",
+    "inaturalist",
+    "biodiversity",
+    "ecology",
+    "invasive-species",
+    "conservation",
+    "species-monitoring"
+  ]
+}
